Respect Maximum Number Of Attempts for the tasks
In this commit, task scheduling parts has been modified which
enforces the controller to respect maximum number of attempts for
the tasks.
Also, it has been observed that when task being dropped and
scheduled again, max number of attempts is not being respected.
in this commit, further checks is added to not schedule the
tasks again once we reach its maximum number of attempts.