Reduce writes to ZK for workflow and job context
In this commit, a deep copy method is added in the TaskDataCache to
keep track of the initial workflow and job contexts. For updating
the context, a comparision is added in the context which basically
compares new and old versions and not update ZK if there is no change
in the contextx.